+++ ergo-lib/src/ast/expr.rs
@@ -2,13 +2,17 @@ use crate::serialization::op_code::OpCode;
 use crate::types::scontext::SContext;
 use crate::types::stype::SType;
 
+use super::and::And;
 use super::apply::Apply;
 use super::bin_op::BinOp;
 use super::block::BlockValue;
 use super::calc_blake2b256::CalcBlake2b256;
 use super::coll_fold::Fold;
+use super::collection::Collection;
 use super::constant::Constant;
 use super::constant::ConstantPlaceholder;
+use super::constant::TryExtractFrom;
+use super::constant::TryExtractFromError;
 use super::extract_amount::ExtractAmount;
 use super::extract_reg_as::ExtractRegisterAs;
 use super::func_value::FuncValue;
@@ -30,6 +34,8 @@ pub enum Expr {
     Const(Constant),
     /// Placeholder for a constant
     ConstPlaceholder(ConstantPlaceholder),
+    /// Collection declaration (array of expressions of the same type)
+    Collection(Collection),
     /// Predefined functions (global)
     /// Blake2b256 hash calculation
     CalcBlake2b256(CalcBlake2b256),
@@ -53,6 +59,8 @@ pub enum Expr {
     ValUse(ValUse),
     /// Binary operation
     BinOp(BinOp),
+    /// Logical AND
+    And(And),
     /// Option get method
     OptionGet(OptionGet),
     /// Extract register's value (box.RX properties)
@@ -69,24 +77,26 @@ impl Expr {
     /// Code (used in serialization)
     pub fn op_code(&self) -> OpCode {
         match self {
-            Expr::Const(_) => todo!(),
-            Expr::ConstPlaceholder(cp) => cp.op_code(),
-            Expr::GlobalVars(v) => v.op_code(),
-            Expr::MethodCall(v) => v.op_code(),
-            Expr::ProperyCall(v) => v.op_code(),
+            Expr::ConstPlaceholder(op) => op.op_code(),
+            Expr::Collection(op) => op.op_code(),
+            Expr::GlobalVars(op) => op.op_code(),
+            Expr::MethodCall(op) => op.op_code(),
+            Expr::ProperyCall(op) => op.op_code(),
             Expr::Context => OpCode::CONTEXT,
-            Expr::OptionGet(v) => v.op_code(),
-            Expr::ExtractRegisterAs(v) => v.op_code(),
+            Expr::OptionGet(op) => op.op_code(),
+            Expr::ExtractRegisterAs(op) => op.op_code(),
             Expr::BinOp(op) => op.op_code(),
             Expr::BlockValue(op) => op.op_code(),
             Expr::ValUse(op) => op.op_code(),
             Expr::FuncValue(op) => op.op_code(),
+            Expr::Apply(op) => op.op_code(),
             Expr::ValDef(op) => op.op_code(),
             Expr::ExtractAmount(op) => op.op_code(),
             Expr::SelectField(op) => op.op_code(),
             Expr::Fold(op) => op.op_code(),
             Expr::CalcBlake2b256(op) => op.op_code(),
-            _ => todo!("not yet implemented opcode for {0:?}", self),
+            Expr::And(op) => op.op_code(),
+            Expr::Const(_) => panic!("constant does not have op code assigned"),
         }
     }
 
@@ -94,6 +104,7 @@ impl Expr {
     pub fn tpe(&self) -> SType {
         match self {
             Expr::Const(v) => v.tpe.clone(),
+            Expr::Collection(v) => v.tpe(),
             Expr::ConstPlaceholder(v) => v.tpe.clone(),
             Expr::CalcBlake2b256(v) => v.tpe(),
             Expr::Context => SType::SContext(SContext()),
@@ -111,6 +122,7 @@ impl Expr {
             Expr::Fold(v) => v.tpe(),
             Expr::SelectField(v) => v.tpe(),
             Expr::ExtractAmount(v) => v.tpe(),
+            Expr::And(v) => v.tpe(),
         }
     }
 
@@ -147,11 +159,24 @@ impl From<InvalidExprEvalTypeError> for InvalidArgumentError {
     }
 }
 
+impl<T: TryExtractFrom<Constant>> TryExtractFrom<Expr> for T {
+    fn try_extract_from(v: Expr) -> Result<Self, super::constant::TryExtractFromError> {
+        match v {
+            Expr::Const(c) => Ok(T::try_extract_from(c)?),
+            _ => Err(TryExtractFromError(format!(
+                "expected Expr::Const, found {:?}",
+                v
+            ))),
+        }
+    }
+}
+
 #[cfg(test)]
 pub mod tests {
     #![allow(unused_imports)]
     use super::*;
     use crate::sigma_protocol::sigma_boolean::SigmaProp;
+    use proptest::collection::*;
     use proptest::prelude::*;
 
     #[derive(PartialEq, Eq, Debug, Clone)]
@@ -178,6 +203,16 @@ pub mod tests {
         .boxed()
     }
 
+    fn coll_nested_expr(depth: usize, elem_tpe: &SType) -> BoxedStrategy<Expr> {
+        match elem_tpe {
+            SType::SBoolean => vec(bool_nested_expr(depth), 0..10)
+                .prop_map(|items| Collection::new(SType::SBoolean, items).unwrap())
+                .prop_map_into(),
+            _ => todo!(),
+        }
+        .boxed()
+    }
+
     fn any_nested_expr(depth: usize) -> BoxedStrategy<Expr> {
         prop_oneof![bool_nested_expr(depth)]
     }
@@ -186,7 +221,7 @@ pub mod tests {
         match tpe {
             SType::SAny => any_nested_expr(depth),
             SType::SBoolean => bool_nested_expr(depth),
-            // SType::SColl(elem_type) => coll_nested_expr(elem_type, depth),
+            SType::SColl(elem_type) => coll_nested_expr(depth, elem_type.as_ref()),
             _ => todo!(),
         }
         .boxed()
@@ -196,8 +231,12 @@ pub mod tests {
         prop_oneof![Just(GlobalVars::Height.into()),].boxed()
     }
 
+    fn bool_non_nested_expr() -> BoxedStrategy<Expr> {
+        prop_oneof![any_with::<Constant>(SType::SBoolean).prop_map_into()].boxed()
+    }
+
     fn any_non_nested_expr() -> BoxedStrategy<Expr> {
-        prop_oneof![int_non_nested_expr()]
+        prop_oneof![int_non_nested_expr(), bool_non_nested_expr()].boxed()
     }
 
     fn coll_non_nested_expr(elem_tpe: &SType) -> BoxedStrategy<Expr> {
@@ -205,7 +244,10 @@ pub mod tests {
             SType::SByte => any_with::<Constant>(SType::SColl(Box::new(SType::SByte)))
                 .prop_map(Expr::Const)
                 .boxed(),
-            _ => todo!(),
+            SType::SBoolean => any_with::<Constant>(SType::SColl(Box::new(SType::SBoolean)))
+                .prop_map(Expr::Const)
+                .boxed(),
+            _ => todo!("Collection of {0:?} is not yet implemented", elem_tpe),
         }
     }
 
@@ -213,8 +255,9 @@ pub mod tests {
         match tpe {
             SType::SAny => any_non_nested_expr(),
             SType::SInt => int_non_nested_expr(),
+            SType::SBoolean => bool_non_nested_expr(),
             SType::SColl(elem_type) => coll_non_nested_expr(elem_type),
-            _ => todo!(),
+            _ => todo!("{0:?} is not yet implemented", tpe),
         }
     }