+++ ergo-lib/src/ast/value.rs
@@ -40,7 +40,7 @@ impl CollPrim {
 
 /// Collection elements
 #[derive(PartialEq, Eq, Debug, Clone)]
-pub enum Coll {
+pub enum CollKind {
     /// Collection elements stored as a vector of primitive types
     Primitive(CollPrim),
     /// Collection elements stored as a vector of ConstantVals
@@ -52,12 +52,12 @@ pub enum Coll {
     },
 }
 
-impl Coll {
+impl CollKind {
     /// Collection element type
     pub fn elem_tpe(&self) -> &SType {
         match self {
-            cp @ Coll::Primitive(_) => cp.elem_tpe(),
-            Coll::NonPrimitive { elem_tpe, .. } => elem_tpe,
+            cp @ CollKind::Primitive(_) => cp.elem_tpe(),
+            CollKind::NonPrimitive { elem_tpe, .. } => elem_tpe,
         }
     }
 }
@@ -86,7 +86,7 @@ pub enum Value {
     /// AVL tree
     AvlTree,
     /// Collection of values of the same type
-    Coll(Coll),
+    Coll(CollKind),
     /// Tuple (arbitrary type values)
     Tup(TupleItems<Value>),
     /// Transaction(and blockchain) context info
@@ -154,13 +154,13 @@ impl From<ErgoBox> for Value {
 
 impl From<Vec<i8>> for Value {
     fn from(v: Vec<i8>) -> Self {
-        Value::Coll(Coll::Primitive(CollPrim::CollByte(v)))
+        Value::Coll(CollKind::Primitive(CollPrim::CollByte(v)))
     }
 }
 
 impl From<Vec<u8>> for Value {
     fn from(v: Vec<u8>) -> Self {
-        Value::Coll(Coll::Primitive(CollPrim::CollByte(v.as_vec_i8())))
+        Value::Coll(CollKind::Primitive(CollPrim::CollByte(v.as_vec_i8())))
     }
 }
 
@@ -181,13 +181,14 @@ impl StoredNonPrimitive for ErgoBox {}
 impl StoredNonPrimitive for EcPoint {}
 impl StoredNonPrimitive for SigmaProp {}
 impl<T: StoredNonPrimitive> StoredNonPrimitive for Option<T> {}
+impl<T> StoredNonPrimitive for Vec<T> {}
 
 #[impl_for_tuples(2, 4)]
 impl StoredNonPrimitive for Tuple {}
 
 impl<T: LiftIntoSType + StoredNonPrimitive + Into<Value>> From<Vec<T>> for Value {
     fn from(v: Vec<T>) -> Self {
-        Value::Coll(Coll::NonPrimitive {
+        Value::Coll(CollKind::NonPrimitive {
             elem_tpe: T::stype(),
             v: v.into_iter().map(|i| i.into()).collect(),
         })
@@ -290,7 +291,7 @@ impl<T: TryExtractFrom<Value> + StoredNonPrimitive> TryExtractFrom<Value> for Ve
     fn try_extract_from(c: Value) -> Result<Self, TryExtractFromError> {
         match c {
             Value::Coll(coll) => match coll {
-                Coll::NonPrimitive { elem_tpe: _, v } => {
+                CollKind::NonPrimitive { elem_tpe: _, v } => {
                     v.into_iter().map(T::try_extract_from).collect()
                 }
                 _ => Err(TryExtractFromError(format!(
@@ -312,7 +313,7 @@ impl TryExtractFrom<Value> for Vec<i8> {
     fn try_extract_from(v: Value) -> Result<Self, TryExtractFromError> {
         match v {
             Value::Coll(v) => match v {
-                Coll::Primitive(CollPrim::CollByte(bs)) => Ok(bs),
+                CollKind::Primitive(CollPrim::CollByte(bs)) => Ok(bs),
                 _ => Err(TryExtractFromError(format!(
                     "expected {:?}, found {:?}",
                     std::any::type_name::<Self>(),
@@ -332,6 +333,12 @@ impl TryExtractFrom<Value> for Vec<u8> {
     fn try_extract_from(v: Value) -> Result<Self, TryExtractFromError> {
         use crate::util::FromVecI8;
         Vec::<i8>::try_extract_from(v).map(Vec::<u8>::from_vec_i8)
+    }
+}
+
+impl TryExtractFrom<Value> for Value {
+    fn try_extract_from(v: Value) -> Result<Self, TryExtractFromError> {
+        Ok(v)
     }
 }
 
@@ -400,5 +407,30 @@ impl TryExtractFrom<Value> for Tuple {
                 v
             ))),
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn byte_u8_array_into() {
+        let bytes = vec![1u8, 2u8, 3u8];
+        let value: Value = bytes.into();
+        assert!(matches!(
+            value,
+            Value::Coll(CollKind::Primitive(CollPrim::CollByte(_)))
+        ))
+    }
+
+    #[test]
+    fn byte_i8_array_into() {
+        let bytes = vec![1i8, 2i8, 3i8];
+        let value: Value = bytes.into();
+        assert!(matches!(
+            value,
+            Value::Coll(CollKind::Primitive(CollPrim::CollByte(_)))
+        ))
     }
 }